refactor: Extract `Size` and `Shape` so we can re-use it for other components
