+++ src/components/Button/index.tsx
@@ -1,5 +1,1 @@
-import { Size as SizeType, Shape as ShapeType } from './styled';
-
 export { Component as Button } from './component';
-export type Size = SizeType;
-export type Shape = ShapeType;
